Now works in Docker
there was some jank that needed to be fixed such as the comments app
itself not being the thing that creates the table, now it does that.
Also made sure to set all the perms up correctly so this isn't running
in the root directory as root user. I can now move onto the items on the
TODO list